requirements.txt
amqp==2.6.1
asgiref==3.2.3
billiard==3.6.3.0
boto3==1.14.38
botocore==1.17.38
celery==4.4.7
certifi==2019.11.28
chardet==3.0.4
confusable-homoglyphs==3.2.0
Django==3.0.2
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-debug-toolbar==2.2
django-froala-editor==3.2.1
django-js-asset==1.2.2
django-registration==3.1
django-storages==1.9.1
django-timezone-field==4.0
dnspython==2.0.0
docutils==0.15.2
ez-setup==0.9
idna==2.9
jmespath==0.10.0
kombu==4.6.11
model-bakery==1.1.1
Pillow==7.0.0
psycopg2-binary==2.8.5
python-crontab==2.5.1
python-dateutil==2.8.1
python-memcached==1.59
pytz==2019.3
redis==3.5.3
requests==2.23.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.2.4
urllib3==1.25.8
vine==1.3.0
